requirements.txt
# MCP Framework
fastmcp==0.1.0

# Database - Using psycopg (v3) with binary support
psycopg[binary]>=3.2.0
psycopg2-binary>=2.9.9  # Required by pgvector
sqlalchemy>=2.0.35
pgvector>=0.3.0

# Embeddings & AI - Using latest versions
openai>=1.54.0
tiktoken>=0.8.0
cohere>=5.11.0

# Utilities - Using latest compatible versions
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.6.1

# CLI & Dev Tools
click>=8.1.7
rich>=13.9.0
python-json-logger>=2.0.7

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
